style: rename common.metatileentity to common.metatileentities
