+++ src/main/kotlin/com/github/trc/clayium/common/metatileentities/PanDuplicatorMetaTileEntity.kt
@@ -1,4 +1,4 @@
-package com.github.trc.clayium.common.metatileentity
+package com.github.trc.clayium.common.metatileentities
 
 import com.cleanroommc.modularui.api.drawable.IKey
 import com.cleanroommc.modularui.factory.PosGuiData
