+++ src/main/kotlin/com/github/trc/clayium/common/metatileentities/SolarClayFabricatorMetaTileEntity.kt
@@ -1,4 +1,4 @@
-package com.github.trc.clayium.common.metatileentity
+package com.github.trc.clayium.common.metatileentities
 
 import com.cleanroommc.modularui.api.drawable.IDrawable
 import com.cleanroommc.modularui.api.drawable.IKey
